requirements.txt
aiogram==3.14.0
black
pre-commit
python-dotenv
#sqlalchemy aiosqlite for future db
